codegen: Factoring out into modules.
